error_reporting admin_value should not be bypassed by user level
This commit reverts the implementation of error_reporting() to the code used in PHP 5.x. The removed code from the PHP 7.x implementation does not seem to add any functionality but miss the checks required to avoid PHP user code to bypass admin_values.
Also most of the code php 7.x implementation code is a duplication of zend_alter_ini_entry() and just php 5.x use it directly.
https://bugs.php.net/bug.php?id=71340